Map volume pick to coordinates in interaction callbacks, remove similar code from visualization container (commented out for now)

+++ interaction/InteractionCallbacks.h
@@ -41,7 +41,7 @@ private:
 	static void MouseMove(vtkRenderWindowInteractor* rwi, VisualizationContainer* vis, ViewType viewType);
 
 	static int Pick(vtkRenderWindowInteractor* rwi);
-	static void VolumePick(double p[3]);
+	static void VolumePick(int p[3]);
 	static void SlicePick(int p[3]);
 
 	InteractionCallbacks();
